Add unit and instumentation tests for ConvertToMetersPerSecond

+++ Lesson1/Task1/Converter/app/src/main/java/ru/geekbrains/converter/ConvertToMetersPerSecond.java
@@ -3,6 +3,6 @@ package ru.geekbrains.converter;
 public class ConvertToMetersPerSecond implements ConvertTo {
     @Override
     public float Do(float sourceValue) {
-        return (sourceValue * 1000)/3600;
+        return (float) ((sourceValue * 1000) / 3600);
     }
 }
